requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
Flask~=2.0
Flask-Compress==1.10.1
Flask-Minify~=0.39
Flask-SocketIO==4.2.1
gunicorn[gthread]~=20.1.0; sys_platform != 'win32'
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown===3.3.4
MarkupSafe==2.0.1
Pillow~=9.0
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
mutagen==1.45.1
PyPDF2==1.27.5
feedparser~=6.0.8
GitPython~=3.1.26
python-dotenv
stripe
pymongo
